+++ resources/views/js.blade.php
@@ -3,8 +3,8 @@
 <script src="https://cdn.paddle.com/paddle/paddle.js"></script>
 <script type="text/javascript">
     @if (config('cashier.sandbox'))
-        Paddle.Environment.set('sandbox');
+        Lnbits.Environment.set('sandbox');
     @endif
 
-    Paddle.Setup(@json($vendor));
+    Lnbits.Setup(@json($vendor));
 </script>