+++ tests/Unit/CashierTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Unit;
 
-use Laravel\Paddle\Cashier;
+use Bitcoin\Lightning\Lnbits\Cashier;
 use Tests\TestCase;
 
 class CashierTest extends TestCase